+++ FlyLocomotionClass.cs
@@ -25,8 +25,8 @@ namespace PatcherYRpp
 		[FieldOffset(81)] public Bool IsLanding;
 		[FieldOffset(82)] public Bool WasLanding;
 		[FieldOffset(83)] public Bool unknown_bool_53; // rotation_byte_53
-		[FieldOffset(84)] public int unknown_54; // rotation_drword_54
-		[FieldOffset(88)] public int unknown_58; // strength_drword_58
+		[FieldOffset(84)] public double unknown_54; // rotation_drword_54
+		[FieldOffset(88)] public double unknown_58; // strength_drword_58
 		[FieldOffset(92)] public Bool IsElevating; // land_shadow_bool
 		[FieldOffset(93)] public Bool unknown_bool_5D;
 		[FieldOffset(94)] public Bool unknown_bool_5E;
